m7: a container-element argument is bridged to the carrier once, not twice
The new vec-of-parametric-sum-monomorph fixture failed under the seam:

  error: incompatible types when assigning to type 'tur_adt_Opt2__int'
         from type 'long int'
    *__t166 = (int64_t)(intptr_t)(__t165);

Two carrier bridges fired on the same `vec-push!` argument, so the second boxed
the FIRST box's pointer into a fresh aggregate slot.

The two blocks were written to divide by the parameter's nominal kind -- the
by-value-APP seam's own comment says "when `A` resolves to a by-value parametric
app the param kind here is TY_APP, so this seam handles it instead" -- but its
guard admits a TY_TYVAR param too.  The split held only because
adt_app_is_byvalue_product answered false for a multi-variant sum; once it
answers true, `vec-push!`'s `val : A` satisfies both.

Fixed by making the intent explicit rather than re-deriving the division: the
first block records that it bridged, and the app seam defers when it did.

On the DEFAULT path this removes one redundant `(int64_t)(intptr_t)` layer from
two vec-push sites (the second block was applying a cast, not a box, over the
first's) -- data-literal-nested's snapshot regenerated in this change, per the
fixture-churn rule.

This was reachable only because the vec-of ICE used to abort the compile before
codegen got here; the ICE fix uncovered the next layer.  Found by the full seam
run, not the subset harness -- which is the reason the harness header now names
`TUR_SR2_APP_SUM_BYVALUE=1 bash tests/run.sh` as the real check.

default: 2711 passed, 0 failed.  seam: 2711 passed, 0 failed.  sr2-seam: 12/0.
sr4-seam: 24/0.

+++ tests/fixtures/data-literal-nested/expected.c
@@ -7987,7 +7987,7 @@ static int64_t vec_hyof_hymaps() {
                     if (tur_panicking) return ((int64_t)0);
                     tur_adt_Vec__Map__sym__int * _un_unv_1337 = __ps_174;
                     (void)_un_unv_1337;
-                    vec_hypush_ex((int64_t)(intptr_t)((int64_t)(intptr_t)((int64_t)(intptr_t)(_un_unv_1337))), (int64_t)(intptr_t)((int64_t)(intptr_t)((int64_t)(intptr_t)(_un_unvw_1336))));
+                    vec_hypush_ex((int64_t)(intptr_t)((int64_t)(intptr_t)((int64_t)(intptr_t)(_un_unv_1337))), (int64_t)(intptr_t)((int64_t)(intptr_t)(_un_unvw_1336)));
                     tur_adt_Map__sym__int * __t175;
                     tur_adt_Map__sym__int * __t176;
                     {
@@ -8020,7 +8020,7 @@ static int64_t vec_hyof_hymaps() {
                         __t176 = __ps_185;
                     }
                     __t175 = __t176;
-                    vec_hypush_ex((int64_t)(intptr_t)((int64_t)(intptr_t)((int64_t)(intptr_t)(_un_unv_1337))), (int64_t)(intptr_t)((int64_t)(intptr_t)((int64_t)(intptr_t)(__t175))));
+                    vec_hypush_ex((int64_t)(intptr_t)((int64_t)(intptr_t)((int64_t)(intptr_t)(_un_unv_1337))), (int64_t)(intptr_t)((int64_t)(intptr_t)(__t175)));
                     tur_adt_Vec__Map__sym__int * __t186;
                     __t186 = _un_unv_1337;
                     __t173 = __t186;
